Move the trex package under executor
Currently, only the `executor` package uses the `trex` package.
