+++ pkg/internal/checkup/executor/trex/types.go
